mod: 简单做些修改，先引入GKPhotoBrowser，之后自己写个

+++ GZEApp/Section/Detail/View/Movie/GZEMovieCastView.m
@@ -52,7 +52,7 @@ static NSInteger kCastCount = 4;
 - (CGSize)itemSize
 {
     CGFloat width = (int)(SCREEN_WIDTH / kCastCount);
-    CGFloat height = width * 1.5 + 70.f;
+    CGFloat height = (int)(width * 1.5 + 70.f);
     return CGSizeMake(width, height);
 }
